Move pending client session to new file
